基本完成施工流程

+++ app/src/main/java/com/renyu/bledemo/utils/ExcelUtils.java
@@ -1,17 +1,22 @@
 package com.renyu.bledemo.utils;
 
-import android.content.Context;
+import android.os.Environment;
 
 import com.renyu.bledemo.params.AddRequestBean;
+import com.renyu.bledemo.params.CommonParams;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.OutputStream;
-import java.util.List;
+import java.util.ArrayList;
 
+import jxl.Sheet;
 import jxl.Workbook;
+import jxl.read.biff.BiffException;
 import jxl.write.Label;
 import jxl.write.WritableSheet;
 import jxl.write.WritableWorkbook;
@@ -24,39 +29,45 @@ import jxl.write.biff.RowsExceededException;
 
 public class ExcelUtils {
 
-    public static void writeExcel(String filePath, List<AddRequestBean> beanList) {
-        File fileDir=new File(filePath);
-        if (!fileDir.exists()) {
-            fileDir.mkdirs();
-        }
-        File file=new File(fileDir+"/test.xls");
-        if (file.exists()) {
-            file.delete();
-        }
-        try {
-            file.createNewFile();
-        } catch (IOException e) {
-            e.printStackTrace();
+    public static void writeExcel(AddRequestBean bean) {
+        File file=new File(Environment.getExternalStorageDirectory().getPath()+File.separator+CommonParams.WRITEFILE);
+        int rowNumbers=getRowNumbers();
+        if (rowNumbers==0) {
+            if (file.exists()) {
+                file.delete();
+            }
+            try {
+                file.createNewFile();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
         WritableWorkbook workbook=null;
+        WritableSheet writableSheet;
         try {
-            OutputStream os=new FileOutputStream(file);
-            workbook= Workbook.createWorkbook(os);
-            WritableSheet writableSheet=workbook.createSheet("测试结果", 0);
-            String[] title={"SN", "测试结果", "测试时间"};
-            for (int i = 0; i < title.length; i++) {
-                Label label=new Label(i, 0, title[i]);
-                writableSheet.addCell(label);
+            if (rowNumbers==0) {
+                OutputStream os=new FileOutputStream(file);
+                workbook= Workbook.createWorkbook(os);
+                writableSheet=workbook.createSheet("测试结果", 0);
+                String[] title={"SN", "deviceId", "测试结果", "测试时间"};
+                for (int i = 0; i < title.length; i++) {
+                    Label label=new Label(i, 0, title[i]);
+                    writableSheet.addCell(label);
+                }
+                rowNumbers=1;
             }
-            for (int i = 0; i < beanList.size(); i++) {
-                AddRequestBean bean=beanList.get(i);
-                Label label0=new Label(0, i+1, bean.getSn());
-                Label label1=new Label(1, i+1, bean.getTestResult());
-                Label label2=new Label(2, i+1, bean.getTestDate());
-                writableSheet.addCell(label0);
-                writableSheet.addCell(label1);
-                writableSheet.addCell(label2);
+            else {
+                workbook= Workbook.createWorkbook(file, Workbook.getWorkbook(file));
+                writableSheet=workbook.getSheet(0);
             }
+            Label label0=new Label(0, rowNumbers, bean.getSn());
+            Label label1=new Label(1, rowNumbers, bean.getDeviceID());
+            Label label2=new Label(2, rowNumbers, bean.getTestResult());
+            Label label3=new Label(3, rowNumbers, bean.getTestDate());
+            writableSheet.addCell(label0);
+            writableSheet.addCell(label1);
+            writableSheet.addCell(label2);
+            writableSheet.addCell(label3);
             workbook.write();
         } catch (FileNotFoundException e) {
             e.printStackTrace();
@@ -65,6 +76,8 @@ public class ExcelUtils {
         } catch (RowsExceededException e) {
             e.printStackTrace();
         } catch (WriteException e) {
+            e.printStackTrace();
+        } catch (BiffException e) {
             e.printStackTrace();
         } finally {
             if (workbook!=null) {
@@ -75,6 +88,64 @@ public class ExcelUtils {
                 } catch (WriteException e) {
                     e.printStackTrace();
                 }
+            }
+        }
+    }
+
+    public static ArrayList<AddRequestBean> readExcel() {
+        String filePath=Environment.getExternalStorageDirectory().getPath() + File.separator + CommonParams.READFILE;
+        ArrayList<AddRequestBean> beans=new ArrayList<>();
+        File file=new File(filePath);
+        if (!file.exists()) {
+            return beans;
+        }
+        try {
+            InputStream is=new FileInputStream(filePath);
+            Workbook workbook=Workbook.getWorkbook(is);
+            Sheet sheet=workbook.getSheet(0);
+            int rows=sheet.getRows();
+            for (int i=0;i<rows;i++) {
+                // 去除第一行标题名称
+                if (i==0) {
+                    continue;
+                }
+                AddRequestBean bean=new AddRequestBean();
+                bean.setDeviceID(sheet.getCell(1, i).getContents());
+                bean.setMachineName(sheet.getCell(2, i).getContents());
+                bean.setMachineId(sheet.getCell(3, i).getContents());
+                beans.add(bean);
+            }
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        } catch (BiffException e) {
+            e.printStackTrace();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return beans;
+    }
+
+    private static int getRowNumbers() {
+        String filePath=Environment.getExternalStorageDirectory().getPath() + File.separator + CommonParams.WRITEFILE;
+        File file=new File(filePath);
+        if (!file.exists()) {
+            return 0;
+        }
+        else {
+            int rows=0;
+            try {
+                InputStream is = new FileInputStream(filePath);
+                Workbook workbook=Workbook.getWorkbook(is);
+                Sheet sheet=workbook.getSheet(0);
+                rows=sheet.getRows();
+            } catch (FileNotFoundException e) {
+                e.printStackTrace();
+            } catch (BiffException e) {
+                e.printStackTrace();
+            } catch (IOException e) {
+                e.printStackTrace();
+            } finally {
+                return rows;
             }
         }
     }
